requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
urllib3==1.25.9
